+++ lib/src/sdk_env.dart
@@ -66,7 +66,9 @@ class SdkConfig {
         ...environment,
         if (resolvedConfigHomePath != null)
           'XDG_CONFIG_HOME': resolvedConfigHomePath,
-        if (resolvedRootPath != null) 'FLUTTER_ROOT': resolvedRootPath,
+        if (resolvedRootPath != null && isFlutterSdk)
+          'FLUTTER_ROOT': resolvedRootPath,
+        if (resolvedRootPath != null) 'PANA_SDK_ROOT_PATH': resolvedRootPath,
       },
     );
   }