+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Omit `pubspec.yaml` warning about local path dependency in the report.
 - Fix type cast issue when customizing `analysis_options.yaml`.
+- Do not export `FLUTTER_ROOT` environment for Dart-only SDK, and export the `PANA_SDK_ROOT_PATH` for all SDKs.
 
 ## 0.23.5